requirements.txt
fastapi
uvicorn[standard]
firebase-admin
google-cloud-firestore
python-dotenv
pydantic[email]
httpx
openai
python-multipart
cachetools
